update sdk
master version: f5a800341e
1. [common][coex][wlan] add vendor info set api for coex
2. [AmebaDplus][WDG] remove assert for IWDG HW enable
3. [common][wlan] move wifi_send_mgnt and delete wifi_set_indicate_mgnt
4. [common][usb] add external usbh interrupt flag
5. [amebasmart][pm] rm unnecessary test code
6. [amebasmart][pm] use api instead of macro in atf pm lib
7. [amebadplus][sdio] merge init code
8. [common][wlan] change param block to u8 for spi device
9. [amebadplus][wlan:fw][bug fix] modify close rf ADDA state
10. [amebadplus][wpan] revert OT_BIG_ENDIAN modify
11. [amebaliate][sdk] fix daily build B4 error
12. [common][pmc] optimize log
13. [amebadplus][amebalite][amebad][sdk] add dependencies by property
14. [amebalite][bt] Zephyr 2nd cycle bredr porting: 6th add part of A2DP src code flow
15. [amebad2][usb]fix UAC bug for ch6
16. [amebasmart][coex] protect bt initial scan
17. [amebadplus][amebalite][amebasmart][sdk] update release tool
18. [amebalite][amebasmart][sdk] update toolchain version check behavior
19. [amebalite]][amebadplus][sdk][wlan] support cmake building for amebalite
20. [amebadplus][pmc] optimze output log and delete unusedless code
21. [common][wlan] update code for max skb size
22. [common][wlan:fw] move interrupt handle code to IC
23. [common][example] Optimize the captive_portal example
24. [common][wlan][bug fix] handle assoc status fail and fix wifi_user_reconnect example
25. [amebad2][gui] Add Smart gui lib release
26. [amebad2][gui] Add third party config
27. [amebadplus][example] add log printf for i2c example
28. [common][wlan] modify inic header file to fix compile warning
29. [common][coex] protect bt initial link 6 data packet interactions
30. [common][atcmd] adjust atcmd_get_ssl_certificate interface

+++ component/wifi/cfg80211_fullmac/rtl8730e/common/rtw_llhw_event_tx.c
@@ -128,7 +128,7 @@ int llhw_wifi_set_mac_addr(u32 wlan_idx, u8 *addr)
 	return ret;
 }
 
-int llhw_wifi_scan(struct _rtw_scan_param_t *scan_param, u32 ssid_length, u32 block)
+int llhw_wifi_scan(struct _rtw_scan_param_t *scan_param, u32 ssid_length, u8 block)
 {
 	int ret = 0;
 	u32 size;
@@ -1035,9 +1035,9 @@ int wifi_btcoex_bt_hci_notify(uint8_t *pdata, uint16_t len, uint8_t dir)
 {
 	int ret = 0;
 
-        (void) pdata;
-        (void) len;
-        (void) dir;
+	(void) pdata;
+	(void) len;
+	(void) dir;
 
 	dev_err(global_idev.fullmac_dev, "%s: not support\n", __func__);
